Change IDs from int to long

+++ features/merchant/review/src/main/java/com/tokopedia/review/common/presentation/uimodel/ReviewProductUiModel.kt
@@ -1,7 +1,7 @@
 package com.tokopedia.review.common.presentation.uimodel
 
 data class ReviewProductUiModel(
-        var productID: Int? = 0,
+        var productID: Long? = 0,
         var productImageUrl: String? = "",
         var productName: String? = "",
         var variantName: String? = ""
