+++ features/merchant/review/src/main/java/com/tokopedia/review/feature/inbox/history/presentation/util/ReviewHistoryItemListener.kt
@@ -1,5 +1,5 @@
 package com.tokopedia.review.feature.inbox.history.presentation.util
 
 interface ReviewHistoryItemListener {
-    fun trackAttachedImageClicked(productId: Int?, feedbackId: Int?)
+    fun trackAttachedImageClicked(productId: Long?, feedbackId: Long?)
 }
