+++ features/merchant/review/src/main/java/com/tokopedia/review/feature/reviewdetail/view/model/FeedbackUiModel.kt
@@ -6,7 +6,7 @@ import com.tokopedia.review.feature.reviewdetail.view.adapter.SellerReviewDetail
 data class FeedbackUiModel(
         var attachments: List<Attachment> = mutableListOf(),
         var autoReply: Boolean = false,
-        var feedbackID: Int? = 0,
+        var feedbackID: Long? = 0,
         var rating: Int? = 0,
         var replyText: String? = "",
         var replyTime: String? = "",
