+++ features/merchant/review/src/main/java/com/tokopedia/review/feature/reviewlist/view/model/ProductReviewUiModel.kt
@@ -3,7 +3,7 @@ package com.tokopedia.review.feature.reviewlist.view.model
 import com.tokopedia.review.feature.reviewlist.view.adapter.SellerReviewListTypeFactory
 
 data class ProductReviewUiModel(
-        var productID: Int? = 0,
+        var productID: Long? = 0,
         var productImageUrl: String? = "",
         var productName: String? = "",
         var rating: Float? = 0.0F,
