+++ features/merchant/review/src/main/java/com/tokopedia/review/feature/reviewreply/view/model/ProductReplyUiModel.kt
@@ -1,7 +1,7 @@
 package com.tokopedia.review.feature.reviewreply.view.model
 
 data class ProductReplyUiModel(
-        var productID: Int? = 0,
+        var productID: Long? = 0,
         var productImageUrl: String? = "",
         var productName: String? = "",
         var variantName: String? = ""
