+++ features/merchant/review/src/main/java/com/tokopedia/review/feature/reviewreply/view/model/UpdateReplyResponseUiModel.kt
@@ -2,8 +2,8 @@ package com.tokopedia.review.feature.reviewreply.view.model
 
 data class UpdateReplyResponseUiModel (
         var isSuccess: Boolean = false,
-        var feedbackId: Int? = 0,
+        var feedbackId: Long? = 0,
         var responseBy: Int? = 0,
-        var shopId: Int? = 0,
+        var shopId: Long? = 0,
         var responseMessage: String? = ""
 )
